cast to float arg1 custom_schedule.py

+++ custom_schedule.py
@@ -9,7 +9,7 @@ class custom_schedule(tf.keras.optimizers.schedules.LearningRateSchedule):
       self.warmup_steps = warmup_steps
 
    def __call__(self, step):
-      arg1 = tf.math.rsqrt(step)
+      arg1 = tf.math.rsqrt(tf.cast(step, tf.float32))
       arg2 = step * (self.warmup_steps ** -1.5)
       return tf.math.rsqrt(self.d_model) * tf.math.minimum(arg1, arg2)
